Convert everything to ES modules
Because of one (1) library I want to use that is an ES module 🫠

This required adding .js file extensions to all the imports, because apparently ES modules in Node are like that. I also removed all the 'use strict' directives because strict mode is enabled by default in ES modules.